neoscreen/requirements.txt
# NeoScreen — Python ML pipeline dependencies
# Install: pip install -r requirements.txt

# Core vision
opencv-python>=4.8.0
numpy>=1.24.0

# ML / inference
tensorflow>=2.13.0          # training + tflite conversion
tflite-runtime>=2.13.0      # lightweight inference-only (no full TF needed at runtime)

# Augmentation
albumentations>=1.3.0

# PHC referral
twilio>=8.0.0
geopy>=2.4.0
python-dotenv>=1.0.0

# Evaluation
scikit-learn>=1.3.0
scipy>=1.11.0

# Utilities
Pillow>=10.0.0
requests>=2.31.0
